requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
httpx>=0.28
python-dotenv>=1.0
pydantic>=2.6
pydantic-settings>=2.2
openai>=1.12
pytest>=8.0
